split structure, styling bugs, CHANGELOG
