+++ README.md
@@ -387,7 +387,7 @@ Giving the default styles in pseudo-code.
 
 | Prop name                      | Type                                                                                                                                                       | Default                                                                   | Description                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    |
 | ------------------------------ | ---------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| sx                             | `SxProps<Theme>`                                                                                                                                           | `{}`                                                                      | Defines `sx` props, that will be **inserted** into Carousel element                                                                                                                                                                                                                                                                                                                                                                                                                                                            |
+| sx                             | `SxProps<Theme>`                                                                                                                                           | `{}`                                                                      | Defines `sx` props, that will be **inserted** into the Carousel 's root element                                                                                                                                                                                                                                                                                                                                                                                                                                                |
 | className                      | `string`                                                                                                                                                   | ""                                                                        | Defines custom class name(s), that will be **added** to Carousel element                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
 | index                          | `number`                                                                                                                                                   | `0`                                                                       | Defines which child (assuming there are more than 1 children) will be displayed. Next and Previous Buttons as well as Indicators will work normally after the first render. When this prop is updated the carousel will display the chosen child. *Use this prop to programmatically set the active child*. If (index > children.length) then if (strictIndexing) index = last element. index                                                                                                                                  |
 | strictIndexing                 | `boolean`                                                                                                                                                  | `true`                                                                    | Defines whether index can be bigger than children length                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
