+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "react-material-ui-carousel",
-    "version": "3.0.2",
+    "version": "3.1.0",
     "license": "MIT",
     "description": "A Generic, extendible Carousel UI component for React using Material UI",
     "homepage": "https://learus.github.io",
